+++ library/src/main/java/com/appdimens/dynamic/code/auto/DimenAutoSp.kt
@@ -437,7 +437,7 @@ object DimenAutoSp {
      * PT Inicia a cadeia de construção para a dimensão customizada AutoSp a partir de um Float base.
      */
     @JvmStatic
-    fun scaled(initialBaseValue: Float): AutoSp = AutoSp(initialBaseValue.toInt())
+    fun scaled(initialBaseValue: Float): AutoSp = AutoSp(initialBaseValue)
 
     // EN Qualifier-based conditional dynamic scaling for Sp.
     // PT Escalonamento condicional baseado em qualificador para Sp.
