+++ library/src/main/java/com/appdimens/dynamic/code/density/DimenDensitySp.kt
@@ -437,7 +437,7 @@ object DimenDensitySp {
      * PT Inicia a cadeia de construção para a dimensão customizada DensitySp a partir de um Float base.
      */
     @JvmStatic
-    fun scaled(initialBaseValue: Float): DensitySp = DensitySp(initialBaseValue.toInt())
+    fun scaled(initialBaseValue: Float): DensitySp = DensitySp(initialBaseValue)
 
     // EN Qualifier-based conditional dynamic scaling for Sp.
     // PT Escalonamento condicional baseado em qualificador para Sp.
