+++ library/src/main/java/com/appdimens/dynamic/code/diagonal/DimenDiagonalSp.kt
@@ -437,7 +437,7 @@ object DimenDiagonalSp {
      * PT Inicia a cadeia de construção para a dimensão customizada DiagonalSp a partir de um Float base.
      */
     @JvmStatic
-    fun scaled(initialBaseValue: Float): DiagonalSp = DiagonalSp(initialBaseValue.toInt())
+    fun scaled(initialBaseValue: Float): DiagonalSp = DiagonalSp(initialBaseValue)
 
     // EN Qualifier-based conditional dynamic scaling for Sp.
     // PT Escalonamento condicional baseado em qualificador para Sp.
