+++ library/src/main/java/com/appdimens/dynamic/code/fit/DimenFitScaled.kt
@@ -290,7 +290,7 @@ class DimenFitScaled private constructor(
         val dpToUse = foundEntry?.customValue ?: initialBaseDp
         val finalQualifier = foundEntry?.finalQualifierResolver ?: qualifier
 
-        return dpToUse.toInt().toDynamicFitDp(
+        return dpToUse.toFloat().toDynamicFitDp(
             context,
             finalQualifier,
             foundEntry?.inverter ?: Inverter.DEFAULT,
