+++ library/src/main/java/com/appdimens/dynamic/code/fluid/DimenFluidSp.kt
@@ -437,7 +437,7 @@ object DimenFluidSp {
      * PT Inicia a cadeia de construção para a dimensão customizada FluidSp a partir de um Float base.
      */
     @JvmStatic
-    fun scaled(initialBaseValue: Float): FluidSp = FluidSp(initialBaseValue.toInt())
+    fun scaled(initialBaseValue: Float): FluidSp = FluidSp(initialBaseValue)
 
     // EN Qualifier-based conditional dynamic scaling for Sp.
     // PT Escalonamento condicional baseado em qualificador para Sp.
