+++ library/src/main/java/com/appdimens/dynamic/code/interpolated/DimenInterpolatedSp.kt
@@ -437,7 +437,7 @@ object DimenInterpolatedSp {
      * PT Inicia a cadeia de construção para a dimensão customizada InterpolatedSp a partir de um Float base.
      */
     @JvmStatic
-    fun scaled(initialBaseValue: Float): InterpolatedSp = InterpolatedSp(initialBaseValue.toInt())
+    fun scaled(initialBaseValue: Float): InterpolatedSp = InterpolatedSp(initialBaseValue)
 
     // EN Qualifier-based conditional dynamic scaling for Sp.
     // PT Escalonamento condicional baseado em qualificador para Sp.
